+++ transactron/lib/__init__.py
@@ -1,7 +1,7 @@
 from .allocators import *  # noqa: F401
 from .fifo import *  # noqa: F401
 from .connectors import *  # noqa: F401
-from .buttons import *  # noqa: F401
+from .basicio import *  # noqa: F401
 from .adapters import *  # noqa: F401
 from .transformers import *  # noqa: F401
 from .reqres import *  # noqa: F401
